Implementation of Jackson and refactor of code
Realised I had to decouple different classes from "parent" classes and/or remove duplicated data. E.g. Crews no longer hold their team object as a member variable, but instead a teamID. A regatta object no longer holds a list of all the crew members, on top of also holding teams which hold crew members inside them as well. All original objects stored in crews, crewmembers and races are now stored as their respective IDs.

Doing this fixed an infinite recursion error with Jackson when trying to serialise to JSON, it would go through a team and convert that into JSON. When doing that it would try to convert a CrewMember to JSON but CrewMember held a team object, so Jackson would then go into that team again repeat the process.